refactor(native): extract sidecar cache cleanup
